+++ packages/pass/components/Item/List/ItemsListItem.tsx
@@ -4,12 +4,15 @@ import { Link, type LinkProps } from 'react-router-dom';
 
 import { ButtonLike, type ButtonLikeProps } from '@proton/atoms/Button';
 import { Icon, Marks } from '@proton/components/components';
+import { useBulkSelect } from '@proton/pass/components/Bulk/BulkSelectProvider';
 import { IconBox } from '@proton/pass/components/Layout/Icon/IconBox';
-import { OptimisticItemIcon } from '@proton/pass/components/Layout/Icon/ItemIcon';
+import { ItemIcon, ItemIconIndicators, SafeItemIcon } from '@proton/pass/components/Layout/Icon/ItemIcon';
 import { itemTypeToSubThemeClassName } from '@proton/pass/components/Layout/Theme/types';
 import { VaultIcon } from '@proton/pass/components/Vault/VaultIcon';
+import { useBulkInFlight } from '@proton/pass/hooks/useBulkInFlight';
 import { useFeatureFlag } from '@proton/pass/hooks/useFeatureFlag';
 import { matchChunks } from '@proton/pass/lib/search/match-chunks';
+import { isWritableVault } from '@proton/pass/lib/vaults/vault.predicates';
 import { selectShare } from '@proton/pass/store/selectors';
 import type { ItemRevisionWithOptimistic, ShareType } from '@proton/pass/types';
 import { PassFeature } from '@proton/pass/types/api/features';
@@ -32,59 +35,96 @@ const ItemsListItemRaw: FC<Props> = ({ item, search = '', active = false, ...res
     const { heading, subheading } = presentListItem(item);
     const vault = useSelector(selectShare<ShareType.Vault>(shareId));
     const pinningEnabled = useFeatureFlag(PassFeature.PassPinningV1);
+    const bulk = useBulkSelect();
+    const bulkSelected = bulk.isSelected(item);
+    const bulkInFlight = useBulkInFlight(item);
+    const loading = optimistic || bulkInFlight;
+    const writable = (vault && isWritableVault(vault)) ?? false;
 
     return (
-        <ButtonLike
-            as={Link}
-            to="#"
-            className={clsx([
-                'pass-item-list--item interactive-pseudo w-full relative',
-                optimistic && !failed && 'opacity-50',
-                active && 'is-active',
-            ])}
-            color={failed ? 'warning' : 'weak'}
-            shape="ghost"
-            {...rest}
-        >
-            <div className="flex-nowrap flex w-full px-3 py-2 items-center">
-                <OptimisticItemIcon
-                    item={item}
-                    size={5}
-                    className={clsx('mr-3  shrink-0', itemTypeToSubThemeClassName[data.type])}
-                />
-                {pinningEnabled && item.pinned && (
-                    <IconBox
-                        size={2.5}
-                        mode="transparent"
-                        className={clsx(
-                            'pass-item-list--item--pin absolute top-custom left-custom flex items-center justify-center',
-                            itemTypeToSubThemeClassName[data.type]
+        <div className={clsx(bulk.enabled && 'px-1 py-0.5')}>
+            <ButtonLike
+                as={Link}
+                to="#"
+                className={clsx([
+                    'pass-item-list--item interactive-pseudo w-full',
+                    bulk.enabled && 'pass-item-list--item-bulk',
+                    bulkSelected && 'pass-item-list--item-bulk-selected',
+                    bulk.enabled && (!writable || bulkInFlight) && 'pointer-events-none opacity-50',
+                    active && 'is-active',
+                ])}
+                color={failed ? 'warning' : 'weak'}
+                shape="ghost"
+                style={{ '--anime-opacity': loading ? '0.5' : '1' }}
+                {...rest}
+            >
+                <div
+                    className={clsx('flex-nowrap flex w-full items-center', bulk.enabled ? 'px-2 py-1.5' : 'px-3 py-2')}
+                >
+                    <SafeItemIcon
+                        item={item}
+                        size={5}
+                        className={clsx('mr-3 shrink-0 relative', itemTypeToSubThemeClassName[data.type])}
+                        iconClassName={clsx(loading && 'opacity-50')}
+                        renderIndicators={(size) => (
+                            <>
+                                {bulk.enabled ? (
+                                    <ItemIcon
+                                        alt=""
+                                        className={'ui-standard absolute bulk-select-check pass-item-list--checkmark'}
+                                        icon={'checkmark'}
+                                        loadImage={false}
+                                        normColor={false}
+                                        pill
+                                        size={5}
+                                    />
+                                ) : (
+                                    <ItemIconIndicators size={size} loading={loading} error={item.failed} />
+                                )}
+
+                                {pinningEnabled && item.pinned && (
+                                    <IconBox
+                                        size={2.5}
+                                        mode="transparent"
+                                        className={clsx(
+                                            'pass-item-list--item--pin absolute bottom-custom right-custom flex items-center justify-center',
+                                            itemTypeToSubThemeClassName[data.type]
+                                        )}
+                                        style={{ '--bottom-custom': '-6px', '--right-custom': '-6px' }}
+                                    >
+                                        <Icon
+                                            name="pin-angled-filled"
+                                            size={2.75}
+                                            className="absolute-center"
+                                            color="black"
+                                        />
+                                    </IconBox>
+                                )}
+                            </>
                         )}
-                        style={{ '--top-custom': '32px', '--left-custom': '36px' }}
-                    >
-                        <Icon name="pin-angled-filled" size={2.75} className="absolute-center" color="black" />
-                    </IconBox>
-                )}
-                <div className="text-left">
-                    <span className="flex items-center">
-                        {search && (
-                            <VaultIcon size={3} icon={vault?.content.display.icon} className="color-weak mr-1" />
-                        )}
-                        <span className="flex-1 text-ellipsis">
-                            <Marks chunks={matchChunks(heading, search)}>{heading}</Marks>
+                    />
+
+                    <div className={clsx('text-left', loading && !failed && 'opacity-50')}>
+                        <span className="flex items-center">
+                            {search && (
+                                <VaultIcon size={3} icon={vault?.content.display.icon} className="color-weak mr-1" />
+                            )}
+                            <span className="flex-1 text-ellipsis">
+                                <Marks chunks={matchChunks(heading, search)}>{heading}</Marks>
+                            </span>
                         </span>
-                    </span>
-                    <div
-                        className={clsx([
-                            'pass-item-list--subtitle block color-weak text-sm text-ellipsis',
-                            item.data.type === 'note' && isEmptyString(item.data.metadata.note.v) && 'text-italic',
-                        ])}
-                    >
-                        <Marks chunks={matchChunks(subheading, search)}>{subheading}</Marks>
+                        <div
+                            className={clsx([
+                                'pass-item-list--subtitle block color-weak text-sm text-ellipsis',
+                                item.data.type === 'note' && isEmptyString(item.data.metadata.note.v) && 'text-italic',
+                            ])}
+                        >
+                            <Marks chunks={matchChunks(subheading, search)}>{subheading}</Marks>
+                        </div>
                     </div>
                 </div>
-            </div>
-        </ButtonLike>
+            </ButtonLike>
+        </div>
     );
 };